修改serial格式不匹配问题

+++ bird/gugu-model/src/main/java/com/gugu/gugumodel/dao/SeminarDao.java
@@ -222,7 +222,7 @@ public class SeminarDao {
      * @param courseId
      * @return
      */
-    public ArrayList<Integer> getSerial(Long courseId){
+    public ArrayList<Byte> getSerial(Long courseId){
         return seminarMapper.getSerial(courseId);
     }
 }
