+++ bird/gugu-model/src/main/java/com/gugu/gugumodel/dao/TeamDao.java
@@ -157,7 +157,7 @@ public class TeamDao{
      * 获取当前所有的serial
      * @param klassId
      */
-    public ArrayList<Integer> getSerial(Long klassId){
+    public ArrayList<Byte> getSerial(Long klassId){
         return teamMapper.getSerial(klassId);
     }
 }
